Updating Boolean type to be intptr_t so get/set funcitions work better.
jeff


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@920 a2d82226-602c-0410-bba7-e51b83dd10d2

+++ owamp/owamp.h
@@ -313,7 +313,7 @@ typedef enum{
     OWP_CNTRL_UNAVAILABLE_TEMP=5    /* Temporary resource limitation    */
 } OWPAcceptType;
 
-typedef uint32_t   OWPBoolean;
+typedef intptr_t   OWPBoolean;
 typedef uint8_t    OWPSID[16];
 typedef uint8_t    OWPSequence[4];
